Comments, documentation and indentation : part 1
This only changes the indentation of some comments (unindented for some
reason) and of the OMP 'pragma' directives. They are aligned with their
outer-most scope (so, one tab for inside class functions and no tab
elsewhere).

Added constexpr specifiers to the size of the generated point clouds in
mainFIST.cpp.

Also, adds some documentation comments and a type alias for the FIST
procedure, in order to make it easier to read. This will progressively
get more 'readable' (according to my standards) along time.