fmt : fix problems with format parsing
As I feared, parsing backwards from the end of the format string does
not work. We need to forward-parse until the next '}' character.

Since the format specification does not allow for '}' characters in the
formatter, simply searching for the next occurence works. No need to
check if it's escaped or not.