Fix glm::mat python interface & func for point<>
The Point<DIM, T> types are now declared like the GLM types, with a
"generic" function which handles the buffer protocol description.

The order of the array indices for the buffer protocol of the glm::mat
types has been fixed (was previously reversed since I cannot seem to
remember what column-major means).