python interface : add wrapper for timings
The timings logger and timing series statistics structures/classes are
now wrapped in a Python type, allowing access to information about the
last run of SPOT/FIST from Python directly.

The micro_benchmark::TimingsLogger class supports the buffer protocol on
the timings themselves, meaning we can access them directly.

On top of that, the Python module declaration was cleaned up.