arch: spot module name : 'spot' -> '_spot'
This is performed in order to have a pure Python module stub (with a
simple __init__.py) as the 'import-able' module, and have all the
implementation details hidden behind the '_spot' module.

The only contents of the __int__.py file will be a bunch of import
statements from the _spot module.

Also, added a setup.py file in order to compile/install the module as a
pip-compatible module.